* commit 'b66085750ec4e8c4a2abab241ebcbf7edbfe469c':
  AArch64: Use long for pointers in media classes

+++ media/java/android/media/AudioRecord.java
@@ -107,13 +107,13 @@ public class AudioRecord
      * Accessed by native methods: provides access to C++ AudioRecord object
      */
     @SuppressWarnings("unused")
-    private int mNativeRecorderInJavaObj;
+    private long mNativeRecorderInJavaObj;
 
     /**
      * Accessed by native methods: provides access to the callback data.
      */
     @SuppressWarnings("unused")
-    private int mNativeCallbackCookie;
+    private long mNativeCallbackCookie;
 
 
     //---------------------------------------------------------
